requirements.txt
numpy
pillow
opencv-python
PyQt5
pyqtgraph
matplotlib
